+++ fsw/src/sbn_client_utils.h
@@ -57,7 +57,7 @@ int message_entry_point(CFE_SBN_Client_PipeD_t);
 int CFE_SBN_CLIENT_ReadBytes(int, unsigned char *, size_t);
 void invalidate_pipe(CFE_SBN_Client_PipeD_t *);
 size_t write_message(int, char *, size_t);
-uint8 CFE_SBN_Client_GetPipeIdx(CFE_SB_PipeId_t);
+uint32 CFE_SBN_Client_GetPipeIdx(CFE_SB_PipeId_t);
 uint8 CFE_SBN_Client_GetMessageSubscribeIndex(CFE_SB_PipeId_t);
 CFE_SB_MsgId_t CFE_SBN_Client_GetMsgId(CFE_MSG_Message_t *);
 int send_heartbeat(int);
